+++ app/com/lynxanalytics/biggraph/frontend_operations/Operations.scala
@@ -408,3 +408,84 @@ object ScalaUtilities {
     }.toIndexedSeq
   }
 }
+
+object PythonUtilities {
+  import com.lynxanalytics.biggraph.graph_operations.DerivePython
+  import DerivePython._
+
+  private def toSerializableType(pythonType: String) = {
+    pythonType match {
+      case "str" => SerializableType.string
+      case "float" => SerializableType.double
+      case _ => throw new AssertionError(s"Unknown type: $pythonType")
+    }
+  }
+
+  def run(
+    code: String, inputs: Seq[String], outputs: Seq[String],
+    project: com.lynxanalytics.biggraph.controllers.ProjectEditor)(
+    implicit
+    manager: MetaGraphManager): Unit = {
+    // Parse the output list into Fields.
+    val api = Seq("vs", "es", "scalars")
+    val outputDeclaration = raw"(\w+)\.(\w+)\s*:\s*(\w+)".r
+    val outputFields = outputs.map {
+      case outputDeclaration(parent, name, tpe) =>
+        assert(
+          api.contains(parent),
+          s"Invalid output: '$parent.$name'. Valid groups are: " + api.mkString(", "))
+        Field(parent, name, toSerializableType(tpe))
+      case output => throw new AssertionError(
+        s"Output declarations must be formatted like 'vs.my_attr: str'. Got '$output'.")
+    }
+    // Parse the input list into Fields.
+    val existingFields = project.vertexAttributes.map {
+      case (name, attr) => s"vs.$name" -> Field("vs", name, SerializableType(attr.typeTag))
+    }.toMap ++ project.edgeAttributes.map {
+      case (name, attr) => s"es.$name" -> Field("es", name, SerializableType(attr.typeTag))
+    }.toMap ++ project.scalars.map {
+      case (name, s) => s"scalars.$name" -> Field("scalars", name, SerializableType(s.typeTag))
+    }.toMap + {
+      "es.src" -> Field("es", "src", SerializableType.long)
+    } + {
+      "es.dst" -> Field("es", "dst", SerializableType.long)
+    }
+    val inputFields = inputs.map { i =>
+      existingFields.get(i) match {
+        case Some(f) => f
+        case None => throw new AssertionError(
+          s"No available input called '$i'. Available inputs are: " +
+            existingFields.keys.toSeq.sorted.mkString(", "))
+      }
+    }
+    // Run the operation.
+    val op = DerivePython(code, inputFields.toList, outputFields.toList)
+    import Scripting._
+    val builder = InstanceBuilder(op)
+    for ((f, i) <- op.attrFields.zipWithIndex) {
+      val attr = f.parent match {
+        case "vs" => project.vertexAttributes(f.name)
+        case "es" => project.edgeAttributes(f.name)
+      }
+      builder(op.attrs(i), attr)
+    }
+    for (f <- op.edgeParents) {
+      builder(op.ebs(f), project.edgeBundle)
+    }
+    for ((f, i) <- op.scalarFields.zipWithIndex) {
+      builder(op.scalars(i), project.scalars(f.name))
+    }
+    builder.toInstance(manager)
+    val res = builder.result
+    // Save the outputs into the project.
+    for ((f, i) <- res.attrFields.zipWithIndex) {
+      f.parent match {
+        case "vs" => project.newVertexAttribute(f.name, res.attrs(i))
+        case "es" => project.newEdgeAttribute(f.name, res.attrs(i))
+      }
+    }
+    for ((f, i) <- res.scalarFields.zipWithIndex) {
+      project.newScalar(f.name, res.scalars(i))
+    }
+  }
+}